removed panics
