+++ crates/utiles/Cargo.toml
@@ -60,6 +60,7 @@ tower-http = { version = "0.5.1", features = ["trace", "timeout", "add-extension
 tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["fmt", "json", "env-filter", "chrono"] }
 walkdir = "2.4.0"
+fnv = "1.0.7"
 # image = "0.25.1"
 
 [dev-dependencies]
