+++ crates/utiles/src/sqlite/dbpath.rs
@@ -21,20 +21,16 @@ impl std::fmt::Display for DbPath {
 impl DbPath {
     #[must_use]
     pub fn new(fspath: &str) -> Self {
-        let p = PathBuf::from(fspath);
-        let filename = p
-            .file_name()
-            .expect("DbPath::new: invalid path, could not get filename from path");
-        DbPath {
-            fspath: fspath.to_string(),
-            filename: filename
-                .to_str()
-                .expect(
-                    "DbPath::new: invalid path, could not convert filename to string",
-                )
-                .to_string(),
-        }
+        // let p = PathBuf::from(fspath);
+        pathlike2dbpath(fspath).map_or(
+            DbPath {
+                fspath: "unknown".to_string(),
+                filename: "unknown".to_string(),
+            },
+            |a| a,
+        )
     }
+
     #[must_use]
     pub fn memory() -> Self {
         DbPath {
