+++ crates/utiles/src/tests/mod.rs
@@ -1,0 +1,3 @@
+#![cfg(test)]
+#![allow(clippy::unwrap_used)]
+mod core;
